+++ CHANGELOG.md
@@ -84,9 +84,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Move to use Intel ifort 2021.13 at NCCS SLES15, NAS, and GMAO Desktops
     - Move to use Intel MPI at NCCS SLES15 and GMAO Desktops
     - Move to GEOSpyD Min24.4.4 Python 3.11
-  - ESMA_cmake v3.48.0
+  - ESMA_cmake v3.49.0
     - Update `esma_add_fortran_submodules` function
     - Move MPI detection out of FindBaselibs
+    - Add SMOD to submodule generator
 
 ### Fixed